requirements.txt
Django==1.11.7
django-debug-toolbar==1.8
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
Question==0.1
six==1.11.0
sqlparse==0.2.4
